+++ src/components/TraditionalDevotionals.tsx
@@ -1,9 +1,9 @@
-import { TTradDevo } from "@root/types";
 import { Container } from "@src/components/Container";
 import { DevotionalCard } from "@src/components/DevotionalCard";
 import { auth, favoriteTradDevo, unfavoriteTradDevo } from "@src/firebase";
 import { useTradDevos } from "@src/hooks/useTradDevos";
 import useStore from "@src/Store";
+import { TTradDevo } from "@src/types";
 import React, { useEffect, useState } from "react";
 import {
   ActivityIndicator,
