+++ src/screens/PastDevotionals.tsx
@@ -1,9 +1,9 @@
-import { TPersonalDevo } from "@root/types";
 import { Container } from "@src/components/Container";
 import { DevotionalCard } from "@src/components/DevotionalCard";
 import { favoritePersonalDevo, unfavoritePersonalDevo } from "@src/firebase";
 import { usePastDevos } from "@src/hooks/usePastDevos";
 import useStore from "@src/Store";
+import { TPersonalDevo } from "@src/types";
 import React, { useEffect, useState } from "react";
 import {
   ActivityIndicator,
